code restructure - move validate_parameter to utils
